requirements.txt
attrs==23.2.0
autopep8==2.0.4
backports.cached-property==1.0.2
build==0.10.0
CacheControl==0.12.14
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==42.0.5
distlib==0.3.8
dulwich==0.21.7
filelock==3.12.2
html5lib==1.1
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
installer==0.7.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.2
pexpect==4.9.0
pipenv-poetry-migrate==0.3.2
pkginfo==1.9.6
pkgutil-resolve-name==1.3.10
platformdirs==3.11.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
psycopg2==2.9.9
ptyprocess==0.7.0
py4j==0.10.9.5
pycodestyle==2.10.0
pycparser==2.21
pyproject-hooks==1.0.0
pyrsistent==0.19.3
pyspark==3.3.3
python-dotenv==0.21.1
rapidfuzz==3.4.0
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
tomli==2.0.1
tomlkit==0.12.4
trove-classifiers==2024.2.23
typer==0.4.2
typing-extensions==4.7.1
urllib3==1.26.18
virtualenv==20.25.1
webencodings==0.5.1
zipp==3.15.0
